+++ src/index.js
@@ -3,13 +3,13 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import NewProvaider from './context/NewProvaider';
+import NewProvider from './context/NewProvider';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <NewProvaider>
+  <NewProvider>
     <App />
-  </NewProvaider>
+  </NewProvider>
 );
 
 // If you want to start measuring performance in your app, pass a function
